feat: #5 - update register test case  - move setup and clean up fonction inside helper file
